🎨Remove generic helpers altogether
