+++ packages/editor-ui/src/components/ExecutionTime.vue
@@ -5,7 +5,6 @@
 </template>
 
 <script lang="ts">
-import { useGenericHelpers } from '@/composables/useGenericHelpers';
 import { defineComponent } from 'vue';
 
 export default defineComponent({
@@ -23,7 +22,7 @@ export default defineComponent({
 				return '...';
 			}
 			const msPassed = this.nowTime - new Date(this.startTime).getTime();
-			return useGenericHelpers().displayTimer(msPassed);
+			return this.$locale.displayTimer(msPassed);
 		},
 	},
 	mounted() {
