+++ packages/editor-ui/src/mixins/workflowHelpers.ts
@@ -48,7 +48,6 @@ import type {
 import { useMessage } from '@/composables/useMessage';
 import { useToast } from '@/composables/useToast';
 import { useNodeHelpers } from '@/composables/useNodeHelpers';
-import { useGenericHelpers } from '@/composables/useGenericHelpers';
 
 import { get, isEqual } from 'lodash-es';
