+++ packages/editor-ui/src/utils/formatters/dateFormatter.ts
@@ -10,3 +10,12 @@ export const convertToDisplayDateComponents = (
 	const [date, time] = formattedDate.split('#');
 	return { date, time };
 };
+
+export function convertToDisplayDate(fullDate: Date | string | number): { date: string; time: string } {
+	const mask = `d mmm${
+		new Date(fullDate).getFullYear() === new Date().getFullYear() ? '' : ', yyyy'
+	}#HH:MM:ss`;
+	const formattedDate = dateformat(fullDate, mask);
+	const [date, time] = formattedDate.split('#');
+	return { date, time };
+}
